+++ extensions/notion/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Notion Changelog
 
-## [New Action to open the Page after creating it] - 2022-04-29
+## [Improvement to the Create Database Page Command] - 2022-04-29
 
 - Added an Action to open the Notion Page after creating it
+- Added a TextArea to add some content to the page directly